Move the user activity into the user panel.
