+++ apps/studio/components/interfaces/Auth/Users/UserActivity.utils.ts
