+++ apps/studio/components/interfaces/Auth/Users/UserActivityEventDetailSheet.tsx
